Fix reservation endpoint mismatch
- Moved full reservation logic from new-reservations.controller.ts to reservations.controller.ts
- Removed /api/v2/reservations route from routes.ts
- Updated integration tests to use /api/reservations consistently
- Updated controller unit tests (renamed new-reservations.controller.test.ts to reservations.controller.test.ts)
- Removed obsolete new-reservations.controller.ts file
- Updated imports in test files to reference correct controller
- All 132 tests passing (integration and unit)
- README already documents /api/reservations as the official endpoint

Resolves endpoint mismatch by consolidating logic under /api/reservations as documented.
